server/requirements.txt
phe==1.5.0
flask==3
python-dotenv